pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
ddg = "*"
flask = "*"
googletrans = "*"
pydictionary = "*"
pyowm = "*"
python = "^3.7"
bs4 = "^0.0.1"
"discord.py" = "^1.2"
json-store-client = "^1.1"
wiktionaryparser = "^0.0.97"
